+++ frontend/src/metabase-lib/parameters/utils/parameter-source.ts
@@ -6,8 +6,10 @@ import {
   ValuesSourceConfig,
   ValuesSourceType,
 } from "metabase-types/api";
-import { getFields } from "metabase-lib/parameters/utils/parameter-fields";
 import Field from "metabase-lib/metadata/Field";
+import { CUSTOM_SOURCE_PARAMETER_TYPES } from "../constants";
+import { getFields } from "./parameter-fields";
+import { getParameterSubType, getParameterType } from "./parameter-type";
 
 export const getQueryType = (parameter: Parameter): ValuesQueryType => {
   return parameter.values_query_type ?? "list";
@@ -19,6 +21,16 @@ export const getSourceType = (parameter: Parameter): ValuesSourceType => {
 
 export const getSourceConfig = (parameter: Parameter): ValuesSourceConfig => {
   return parameter.values_source_config ?? {};
+};
+
+export const canUseCustomSource = (parameter: Parameter) => {
+  const type = getParameterType(parameter);
+  const subType = getParameterSubType(parameter);
+
+  return (
+    CUSTOM_SOURCE_PARAMETER_TYPES[type] != null &&
+    CUSTOM_SOURCE_PARAMETER_TYPES[type].includes(subType)
+  );
 };
 
 export const isValidSourceConfig = (
