+++ frontend/src/metabase/parameters/utils/parameter-type.ts
@@ -1,11 +1,8 @@
 import { Parameter } from "metabase-types/api";
+import { SINGLE_OR_MULTI_SELECTABLE_TYPES } from "metabase-lib/parameters/constants";
 import {
-  CUSTOM_SOURCE_PARAMETER_TYPES,
-  SINGLE_OR_MULTI_SELECTABLE_TYPES,
-} from "metabase-lib/parameters/constants";
-import {
-  getParameterType,
   getParameterSubType,
+  getParameterType,
 } from "metabase-lib/parameters/utils/parameter-type";
 
 export function isSingleOrMultiSelectable(parameter: Parameter): boolean {
@@ -20,13 +17,3 @@ export function isSingleOrMultiSelectable(parameter: Parameter): boolean {
   }
   return SINGLE_OR_MULTI_SELECTABLE_TYPES[type].includes(subType);
 }
-
-export const canUseCustomSource = (parameter: Parameter) => {
-  const type = getParameterType(parameter);
-  const subType = getParameterSubType(parameter);
-
-  return (
-    CUSTOM_SOURCE_PARAMETER_TYPES[type] != null &&
-    CUSTOM_SOURCE_PARAMETER_TYPES[type].includes(subType)
-  );
-};
