+++ frontend/src/metabase/query_builder/components/template_tags/TagEditorParam.jsx
@@ -12,13 +12,14 @@ import InputBlurChange from "metabase/components/InputBlurChange";
 import Select, { Option } from "metabase/core/components/Select";
 
 import ValuesSourceSettings from "metabase/parameters/components/ValuesSourceSettings";
-import { canUseCustomSource } from "metabase/parameters/utils/parameter-type";
 import { getParameterOptionsForField } from "metabase/parameters/utils/template-tag-options";
 
 import { fetchField } from "metabase/redux/metadata";
 import { getMetadata } from "metabase/selectors/metadata";
 import { SchemaTableAndFieldDataSelector } from "metabase/query_builder/components/DataSelector";
 import MetabaseSettings from "metabase/lib/settings";
+
+import { canUseCustomSource } from "metabase-lib/parameters/utils/parameter-source";
 
 import {
   ErrorSpan,
